integrate datasets
